+++ permission/sample/src/main/java/ru/surfstudio/android/permission/sample/screen/main/location/LocationPermissionRequest.kt
@@ -1,0 +1,13 @@
+package ru.surfstudio.android.permission.sample.screen.main.location
+
+import android.Manifest
+import ru.surfstudio.android.core.ui.permission.PermissionRequest
+
+class LocationPermissionRequest : PermissionRequest() {
+
+    override val permissions: Array<String>
+        get() = arrayOf(
+            Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.ACCESS_COARSE_LOCATION
+        )
+}