+++ push/lib-push/src/main/java/ru/surfstudio/android/notification/ui/notification/strategies/PushHandleStrategy.kt
@@ -43,11 +43,16 @@ import java.io.Serializable
 abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> : Serializable {
 
     /**
-     * Идентификатор пуш-нотификации. При совпадении идентификаторов пуш-нотификации заменяют друг
-     * друга. Если в стратегии не переопределить данное свойство, все нотификации будут
-     * автоматически получать уникальные идентификаторы.
+     * Идентификатор пуш-нотификации.
+     * При совпадении идентификаторов пуш-нотификации заменяют друг друга.
      */
     open var pushId: Int = -1
+
+    /**
+     * Признак уникальности пушей. Если в стратегии не переопределить данное свойство, все нотификации будут
+     * автоматически получать уникальные идентификаторы.
+     */
+    open var uniquePushes: Boolean = true
 
     /**
      * тип данных пуша [BaseNotificationTypeData]
@@ -133,7 +138,11 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> : Seriali
      * @param title заголовок пуш-нотификации
      * @param body текст пуш-нотификации
      */
-    abstract fun makeNotificationBuilder(context: Context, title: String, body: String): NotificationCompat.Builder?
+    abstract fun makeNotificationBuilder(
+        context: Context,
+        title: String,
+        body: String
+    ): NotificationCompat.Builder?
 
     /**
      * Метод для создания канала нотификаций
@@ -150,14 +159,15 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> : Seriali
      * @param body текст пуш-нотификации
      */
     open fun handle(
-            context: Context,
-            pushInteractor: PushInteractor,
-            uniqueId: Int,
-            title: String,
-            body: String
+        context: Context,
+        pushInteractor: PushInteractor,
+        uniqueId: Int,
+        title: String,
+        body: String
     ) {
-
-        pushId = makePushId(uniqueId)
+        if (uniquePushes) {
+            pushId = uniqueId
+        }
         pendingIntent = preparePendingIntent(context, title, group?.id, pushId)
         deleteIntent = makeDeleteIntent(context, group?.id, pushId)
         channel = makeNotificationChannel(context, title)
@@ -178,13 +188,15 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> : Seriali
      * @param title заголовок пуш-нотификации
      * @param body текст пуш-нотификации
      */
-    @Deprecated("Используйте метод с uniqueId",
-            ReplaceWith("handle(context, pushInteractor, uniqueId, title, body"))
+    @Deprecated(
+        "Используйте метод с uniqueId",
+        ReplaceWith("handle(context, pushInteractor, uniqueId, title, body")
+    )
     open fun handle(
-            context: Context,
-            pushInteractor: PushInteractor,
-            title: String,
-            body: String
+        context: Context,
+        pushInteractor: PushInteractor,
+        title: String,
+        body: String
     ) {
 
         pendingIntent = preparePendingIntent(context, title, -1, -1)
@@ -200,9 +212,11 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> : Seriali
     /**
      * Метод для инициализации builder'а заголовка группы нотификаций.
      */
-    open fun makeGroupSummaryNotificationBuilder(context: Context,
-                                                 title: String,
-                                                 body: String): NotificationCompat.Builder? {
+    open fun makeGroupSummaryNotificationBuilder(
+        context: Context,
+        title: String,
+        body: String
+    ): NotificationCompat.Builder? {
         return null
     }
 
@@ -217,60 +231,68 @@ abstract class PushHandleStrategy<out T : BaseNotificationTypeData<*>> : Seriali
     }
 
     @CallSuper
-    protected open fun getIntentForPending(context: Context, groupId: Int?, notificationId: Int?): Intent {
+    protected open fun getIntentForPending(
+        context: Context,
+        groupId: Int?,
+        notificationId: Int?
+    ): Intent {
         return Intent(context, NotificationClickEventReceiver::class.java)
-                .apply {
-                    putExtra(NOTIFICATION_DATA, typeData)
-                    putExtra(NOTIFICATION_GROUP_ID, groupId ?: 0)
-                    putExtra(NOTIFICATION_ID, notificationId ?: 0)
-                    putExtra(SHOULD_AUTO_CANCEL, autoCancelable)
-                }
+            .apply {
+                putExtra(NOTIFICATION_DATA, typeData)
+                putExtra(NOTIFICATION_GROUP_ID, groupId ?: 0)
+                putExtra(NOTIFICATION_ID, notificationId ?: 0)
+                putExtra(SHOULD_AUTO_CANCEL, autoCancelable)
+            }
     }
 
     /**
      * Интент в соответствии с необходимыми действиями
      */
-    private fun preparePendingIntent(context: Context, title: String, groupId: Int?, notificationId: Int?): PendingIntent {
+    private fun preparePendingIntent(
+        context: Context,
+        title: String,
+        groupId: Int?,
+        notificationId: Int?
+    ): PendingIntent {
         val intent = getIntentForPending(context, groupId, notificationId).apply {
             putExtra(EVENT_TYPE, Event.OPEN)
         }
-        return PendingIntent.getBroadcast(context.applicationContext,
-                title.hashCode(), intent, PendingIntent.FLAG_ONE_SHOT)
+        return PendingIntent.getBroadcast(
+            context.applicationContext,
+            title.hashCode(), intent, PendingIntent.FLAG_ONE_SHOT
+        )
     }
 
-    private fun makeDeleteIntent(context: Context, groupId: Int?, notificationId: Int?): PendingIntent {
+    private fun makeDeleteIntent(
+        context: Context,
+        groupId: Int?,
+        notificationId: Int?
+    ): PendingIntent {
         val intent = getIntentForPending(context, groupId, notificationId).apply {
             putExtra(EVENT_TYPE, Event.DISMISS)
         }
-        return PendingIntent.getBroadcast(context.applicationContext,
-                0, intent, PendingIntent.FLAG_ONE_SHOT)
-    }
-
-    /**
-     * Инициализация идентификатора пуш-нотификации.
-     *
-     * Если ID пуша не задано явно в стратегии, все пуши будут иметь уникальный ID.
-     */
-    private fun makePushId(uniqueId: Int): Int {
-        return if (pushId == -1) uniqueId else pushId
+        return PendingIntent.getBroadcast(
+            context.applicationContext,
+            0, intent, PendingIntent.FLAG_ONE_SHOT
+        )
     }
 
     private fun showNotification(context: Context, pushId: Int, title: String, body: String) {
         NotificationCreateHelper.showNotification(
-                context,
-                this,
-                pushId,
-                title,
-                body
+            context,
+            this,
+            pushId,
+            title,
+            body
         )
     }
 
     private fun showNotification(context: Context, title: String, body: String) {
         NotificationCreateHelper.showNotification(
-                context,
-                this,
-                title,
-                body
+            context,
+            this,
+            title,
+            body
         )
     }
 }