+++ template/base_feature/src/main/java/ru/surfstudio/standard/application/app/App.kt
@@ -31,17 +31,17 @@ class App : Application() {
 
         initAnrWatchDog()
         initLog()
-        registerActiveActivityListener()
 
         RxJavaPlugins.setErrorHandler { Logger.e(it) }
         AppInjector.initInjector(this)
         DebugAppInjector.initInjector(this, activeActivityHolder)
+        registerActiveActivityListener()
 
         //initFirebaseCrashlytics() todo uncoment for a real app
         initPushEventListener()
         initRxJava2Debug()
         registerNavigationProviderCallbacks()
-        DebugAppInjector.debugInteractor.onCreateApp(R.mipmap.ic_launcher)
+        DebugAppInjector.debugInteractor.onCreateApp(R.drawable.ic_android)
     }
 
     private fun registerNavigationProviderCallbacks() {
@@ -55,8 +55,8 @@ class App : Application() {
      */
     private fun initAnrWatchDog() {
         ANRWatchDog().setReportMainThreadOnly()
-                .setANRListener { RemoteLogger.logError(it) }
-                .start()
+            .setANRListener { RemoteLogger.logError(it) }
+            .start()
     }
 
     private fun initLog() {
@@ -73,15 +73,16 @@ class App : Application() {
      * Регистрирует слушатель аткивной активити
      */
     private fun registerActiveActivityListener() {
+        registerActivityLifecycleCallbacks(AppInjector.appComponent.navigationCallbacks())
         registerActivityLifecycleCallbacks(
-                ActivityLifecycleListener(
-                        onActivityResumed = { activity ->
-                            activeActivityHolder.activity = activity
-                        },
-                        onActivityPaused = {
-                            activeActivityHolder.clearActivity()
-                        }
-                )
+            ActivityLifecycleListener(
+                onActivityResumed = { activity ->
+                    activeActivityHolder.activity = activity
+                },
+                onActivityPaused = {
+                    activeActivityHolder.clearActivity()
+                }
+            )
         )
     }
 
@@ -94,7 +95,7 @@ class App : Application() {
     private fun initPushEventListener() {
         PushClickProvider.pushEventListener = object : PushEventListener {
             override fun pushDismissListener(context: Context, intent: Intent) {
-                //todo
+                /* do nothing */
             }
 
             override fun pushOpenListener(context: Context, intent: Intent) {
@@ -102,7 +103,7 @@ class App : Application() {
             }
 
             override fun customActionListener(context: Context, intent: Intent) {
-                //todo
+                /* do nothing */
             }
         }
     }